+++ includes/i18n/el.php
@@ -28,6 +28,7 @@ $i18n = [
     'reset_sent_check_email' => "Ένα email με οδηγίες για την επαναφορά του κωδικού πρόσβασης σας έχει σταλεί. Παρακαλώ ελέγξτε το email σας.",
     'password_reset_successful' => "Επιτυχής επαναφορά κωδικού πρόσβασης",
     // Header
+    'profile' => "Προφίλ",
     'subscriptions' => "Συνδρομές",
     'stats' => "Στατιστικές",
     'settings' => "Ρυθμίσεις",
@@ -116,7 +117,7 @@ $i18n = [
     'the_author' => "Προγραμματιστής",
     'icons' => "Εικονίδια",
     'payment_icons' => "Εικονίδια Payment",
-    // Settings page
+    // Profile page
     'upload_avatar' => "μεταφόρτωση άβαταρ",
     'file_type_error' => "Το αρχείο πρέπει να είναι τύπου jpg, jpeg, png, webp ή gif",
     'user_details' => "Λεπτομέρειες χρήστη",
@@ -132,6 +133,10 @@ $i18n = [
     "totp_backup_codes_info" => "Αποθηκεύστε αυτούς τους κωδικούς ανάκτησης σε ένα ασφαλές μέρος. Θα χρειαστείτε έναν από αυτούς τους κωδικούς ανάκτησης για να αποκτήσετε πρόσβαση στο λογαριασμό σας σε περίπτωση που χάσετε τη συσκευή σας.",
     "disable_two_factor_authentication" => "Απενεργοποίηση διπλής πιστοποίησης",
     "totp_code" => "Κωδικός TOTP",
+    "api_key" => "API κλειδί",
+    "regenerate" => "Επαναδημιουργία",
+    "api_key_info" => "Το API κλειδί χρησιμοποιείται για την επικοινωνία με το API του Wallos. Μην αποκαλύπτετε το API κλειδί σας σε κανέναν.",
+    // Settings page
     "monthly_budget" => "Μηνιαίος προϋπολογισμός",
     "budget_info" => "Ο μηνιαίος προϋπολογισμός χρησιμοποιείται για τον υπολογισμό των στατιστικών",
     "household" => "Νοικοκυριό",
