+++ includes/i18n/sr.php
@@ -28,6 +28,7 @@ $i18n = [
     'reset_sent_check_email' => "Ресетовање лозинке је послато на вашу е-пошту",
     'password_reset_successful' => "Ресетовање лозинке је успешно",
     // Header
+    'profile' => "Профил",
     'subscriptions' => "Претплате",
     'stats' => "Статистике",
     'settings' => "Подешавања",
@@ -116,7 +117,7 @@ $i18n = [
     'the_author' => "Аутор",
     'icons' => "Иконе",
     'payment_icons' => "Иконе плаћања",
-    // Страница са подешавањима
+    // Profile page
     'upload_avatar' => "Постави аватар",
     'file_type_error' => "Датотека није у подржаном формату.",
     'user_details' => "Кориснички детаљи",
@@ -132,6 +133,10 @@ $i18n = [
     "totp_backup_codes_info" => "Сачувајте ове кодове на безбедно место. Користићете их када изгубите приступ апликацији за аутентификацију.",
     "disable_two_factor_authentication" => "Онемогући двофакторску аутентикацију",
     "totp_code" => "ТОТП код",
+    "api_key" => "API кључ",
+    "regenerate" => "Генериши",
+    "api_key_info" => "API кључ је потребан за коришћење API-ја за приступ вашим подацима. Не делите овај кључ са другима.",
+    // Страница са подешавањима
     "monthly_budget" => "Месечни буџет",
     "budget_info" => "Унесите месечни буџет да бисте видели препоручену максималну цену претплате на почетној страници.",
     "household" => "Домаћинство",
